One-type-per-file split of CmuxMobileShellUI; dedup keyboard dismissal (iOS refactor 5079 wave 3)
Fan out the moved multi-type view files into one major type per file per the
package file-organization rule:
- WorkspaceViews.swift -> WorkspaceShellView / WorkspaceNavigationStyle /
  WorkspaceDetailContainer / WorkspaceListView / WorkspaceNavigationRow /
  WorkspaceRow / WorkspaceDetailView / MobileWorkspacePreview+Display.
- SignInView.swift -> SignInView / RestoringSessionView / DividerLabel /
  GameOfLifeHeader / GameOfLifeGrid / GlassInputPill.
- PairingView.swift -> PairingView / AddDeviceInputKind / MobilePairingScannerSheet /
  QRCodeScannerView / MobilePairingScannerPredicate.
- CMUXMobileFeature.swift -> CMUXMobileAppView / CMUXMobileRootView /
  DisconnectedWorkspaceShellView / MobileRootAuthGate+ShellSync.
- MobileTerminalSafeArea.swift -> TerminalPalette / PlatformPalette /
  View+MobileTerminalSafeArea / View+MobileTextInput / View+MobileNavigationChrome /
  View+MobileGlass.

Dedups the two copy-pasted dismissKeyboard() helpers into one shared
dismissMobileKeyboard(). Deletes the dead MobileTerminalDeviceSafeArea
namespace-enum and the unused mobileShellUILog. Cross-file private types are
promoted to package-internal.

Extracts CMUXMobileRootView's content switch into a rootContent property so the
body modifier chain type-checks within the solver budget after the split.

Marks the remaining TerminalAccessoryConfiguration.shared view reach-in
TRANSITIONAL: it is shared with the off-limits typing-latency render path, so its
de-singletonization is deferred to the terminal-surface wave. Behavior unchanged.
