+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/AddDeviceInputKind.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+/// The keyboard / content-type behavior for an add-device text field.
+enum AddDeviceInputKind {
+    case text
+    case url
+    case number
+}
