+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/CMUXMobileAppView.swift
@@ -1,0 +1,27 @@
+import Foundation
+import CMUXMobileCore
+import CmuxAuthRuntime
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import Observation
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+public struct CMUXMobileAppView: View {
+    @State private var store: CMUXMobileShellStore
+
+    public init(store: CMUXMobileShellStore = .preview()) {
+        _store = State(initialValue: store)
+    }
+
+    public var body: some View {
+        CMUXMobileRootView(store: store)
+    }
+}
