+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/DividerLabel.swift
@@ -1,0 +1,37 @@
+import Foundation
+import CMUXMobileCore
+import CmuxAuthRuntime
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import StackAuth
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct DividerLabel: View {
+    let text: String
+
+    var body: some View {
+        HStack(spacing: 8) {
+            dividerLine
+            Text(text)
+                .font(.caption2)
+                .foregroundStyle(Color.primary.opacity(0.45))
+                .lineLimit(1)
+                .minimumScaleFactor(0.6)
+                .allowsTightening(true)
+                .layoutPriority(1)
+            dividerLine
+        }
+    }
+
+    private var dividerLine: some View {
+        Rectangle()
+            .fill(PlatformPalette.separator.opacity(0.4))
+            .frame(height: 1)
+    }
+}
