+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/GameOfLifeGrid.swift
@@ -1,0 +1,136 @@
+import Foundation
+import CMUXMobileCore
+import CmuxAuthRuntime
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import StackAuth
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct GameOfLifeGrid: View {
+    let columns: Int
+    let rows: Int
+
+    @State private var cells: [Bool] = []
+    @State private var stepCount = 0
+    @Environment(\.colorScheme) private var colorScheme
+    @Environment(\.displayScale) private var displayScale
+
+    var body: some View {
+        TimelineView(.periodic(from: .now, by: 0.08)) { timeline in
+            let time = timeline.date.timeIntervalSinceReferenceDate
+            let tick = Int(time / 0.08)
+
+            GeometryReader { _ in
+                Canvas { context, size in
+                    let cellWidth = size.width / CGFloat(columns)
+                    let cellHeight = size.height / CGFloat(rows)
+                    let cellSize = min(cellWidth, cellHeight) * 0.52
+                    let yOffset = (cellHeight - cellSize) * 0.5
+                    let xOffset = (cellWidth - cellSize) * 0.5
+                    let scale = max(1, displayScale)
+
+                    func snapToPixel(_ value: CGFloat) -> CGFloat {
+                        (value * scale).rounded(.toNearestOrAwayFromZero) / scale
+                    }
+
+                    for row in 0..<rows {
+                        for col in 0..<columns where isAlive(row: row, col: col) {
+                            let baseOpacity = colorScheme == .dark ? 0.10 : 0.16
+                            let flicker = baseOpacity + 0.10 * sin(time * 2.0 + Double(row * 3 + col) * 0.22)
+                            let rect = CGRect(
+                                x: snapToPixel(CGFloat(col) * cellWidth + xOffset),
+                                y: snapToPixel(CGFloat(row) * cellHeight + yOffset),
+                                width: snapToPixel(cellSize),
+                                height: snapToPixel(cellSize)
+                            )
+                            context.fill(
+                                Path(roundedRect: rect, cornerRadius: rect.width * 0.5),
+                                with: .color(PlatformPalette.gameOfLifeCell(colorScheme: colorScheme).opacity(max(0.0, flicker)))
+                            )
+                        }
+                    }
+                }
+            }
+            .onChange(of: tick) { _, _ in
+                step()
+            }
+            .onAppear {
+                if cells.isEmpty {
+                    seed()
+                }
+            }
+        }
+    }
+
+    private func index(row: Int, col: Int) -> Int {
+        row * columns + col
+    }
+
+    private func isAlive(row: Int, col: Int) -> Bool {
+        let wrappedRow = (row + rows) % rows
+        let wrappedCol = (col + columns) % columns
+        let idx = index(row: wrappedRow, col: wrappedCol)
+        if idx < cells.count {
+            return cells[idx]
+        }
+        return false
+    }
+
+    private func seed() {
+        var rng = SystemRandomNumberGenerator()
+        cells = (0..<(rows * columns)).map { _ in
+            Double.random(in: 0...1, using: &rng) < 0.22
+        }
+        stepCount = 0
+    }
+
+    private func step() {
+        guard !cells.isEmpty else {
+            seed()
+            return
+        }
+
+        var next = cells
+        var aliveCount = 0
+
+        for row in 0..<rows {
+            for col in 0..<columns {
+                let idx = index(row: row, col: col)
+                let neighbors = neighborCount(row: row, col: col)
+                let alive = cells[idx]
+                let nextAlive = (alive && (neighbors == 2 || neighbors == 3)) || (!alive && neighbors == 3)
+                next[idx] = nextAlive
+                if nextAlive {
+                    aliveCount += 1
+                }
+            }
+        }
+
+        stepCount += 1
+
+        if aliveCount < max(6, (rows * columns) / 22) || stepCount > 120 {
+            seed()
+            return
+        }
+
+        cells = next
+    }
+
+    private func neighborCount(row: Int, col: Int) -> Int {
+        var count = 0
+        for dr in -1...1 {
+            for dc in -1...1 where dr != 0 || dc != 0 {
+                if isAlive(row: row + dr, col: col + dc) {
+                    count += 1
+                }
+            }
+        }
+        return count
+    }
+}
