+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/GameOfLifeHeader.swift
@@ -1,0 +1,40 @@
+import Foundation
+import CMUXMobileCore
+import CmuxAuthRuntime
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import StackAuth
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct GameOfLifeHeader: View {
+    private let columns = 36
+    private let rows = 52
+    @Environment(\.colorScheme) private var colorScheme
+
+    var body: some View {
+        GeometryReader { proxy in
+            ZStack {
+                GameOfLifeGrid(columns: columns, rows: rows)
+                    .padding(.horizontal, 6)
+                    .padding(.vertical, 6)
+
+                LinearGradient(
+                    colors: [
+                        PlatformPalette.systemBackground.opacity(0.0),
+                        PlatformPalette.systemBackground.opacity(colorScheme == .dark ? 0.82 : 0.70),
+                    ],
+                    startPoint: .center,
+                    endPoint: .bottom
+                )
+            }
+            .frame(width: proxy.size.width, height: proxy.size.height)
+        }
+        .clipped()
+    }
+}
