+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/MobilePairingScannerSheet.swift
@@ -1,0 +1,68 @@
+import CMUXMobileCore
+import CmuxMobileCamera
+import CmuxMobileSupport
+import SwiftUI
+#if os(iOS)
+import UIKit
+#endif
+
+#if os(iOS)
+struct MobilePairingScannerSheet: View {
+    let onCode: (String) -> Void
+    @Environment(\.dismiss) private var dismiss
+    private let authorization = CameraAuthorization()
+    @State private var authorizationStatus = CameraAuthorization().videoStatus
+
+    var body: some View {
+        NavigationStack {
+            Group {
+                switch authorizationStatus {
+                case .authorized:
+                    QRCodeScannerView { code in
+                        onCode(code)
+                    }
+                    .ignoresSafeArea(edges: .bottom)
+                case .notDetermined:
+                    ProgressView()
+                        .task {
+                            authorizationStatus = await authorization.requestVideoAccess()
+                        }
+                case .denied, .restricted:
+                    ContentUnavailableView(
+                        L10n.string("mobile.pairing.cameraDenied", defaultValue: "Camera Access Required"),
+                        systemImage: "camera.fill"
+                    )
+                @unknown default:
+                    ContentUnavailableView(
+                        L10n.string("mobile.pairing.cameraUnavailable", defaultValue: "Camera Unavailable"),
+                        systemImage: "camera.fill"
+                    )
+                }
+            }
+            .navigationTitle(L10n.string("mobile.pairing.scannerTitle", defaultValue: "Scan QR Code"))
+            .navigationBarTitleDisplayMode(.inline)
+            .toolbar {
+                ToolbarItem(placement: .topBarTrailing) {
+                    Button {
+                        dismiss()
+                    } label: {
+                        Text(L10n.string("mobile.pairing.scannerCancel", defaultValue: "Cancel"))
+                    }
+                    .accessibilityIdentifier("MobileScannerCancelButton")
+                }
+            }
+        }
+    }
+}
+#else
+struct MobilePairingScannerSheet: View {
+    let onCode: (String) -> Void
+
+    var body: some View {
+        ContentUnavailableView(
+            L10n.string("mobile.pairing.cameraUnavailable", defaultValue: "Camera Unavailable"),
+            systemImage: "camera.fill"
+        )
+    }
+}
+#endif
