--- Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/MobileTerminalSafeArea.swift
@@ -1,290 +1,0 @@
-import Foundation
-@preconcurrency import AVFoundation
-import CMUXMobileCore
-import CmuxMobileTerminal
-import CmuxMobileWorkspace
-import Observation
-import OSLog
-import StackAuth
-import SwiftUI
-#if os(iOS)
-@preconcurrency import UIKit
-#elseif os(macOS)
-import AppKit
-#endif
-
-struct TerminalPalette {
-    private init() {}
-
-    static let background = Color(red: 0x27 / 255.0, green: 0x28 / 255.0, blue: 0x22 / 255.0)
-    static let foreground = Color(red: 0xf8 / 255.0, green: 0xf8 / 255.0, blue: 0xf2 / 255.0)
-    static let dimForeground = Color(red: 0xc8 / 255.0, green: 0xc8 / 255.0, blue: 0xc0 / 255.0)
-}
-
-struct PlatformPalette {
-    private init() {}
-
-    static var systemBackground: Color {
-        #if os(iOS)
-        Color(uiColor: .systemBackground)
-        #elseif os(macOS)
-        Color(nsColor: .windowBackgroundColor)
-        #else
-        Color.white
-        #endif
-    }
-
-    static var separator: Color {
-        #if os(iOS)
-        Color(uiColor: .separator)
-        #elseif os(macOS)
-        Color(nsColor: .separatorColor)
-        #else
-        Color.gray
-        #endif
-    }
-
-    static func gameOfLifeCell(colorScheme: ColorScheme) -> Color {
-        #if os(iOS)
-        Color(uiColor: colorScheme == .dark ? .systemGray4 : .systemGray2)
-        #elseif os(macOS)
-        Color(nsColor: colorScheme == .dark ? .systemGray : .secondaryLabelColor)
-        #else
-        Color.gray
-        #endif
-    }
-}
-
-
-#if os(iOS)
-private struct MobileCompactLandscapeTerminalSafeAreaCompensation: ViewModifier {
-    let context: MobileTerminalSafeAreaContext
-    let includesBottom: Bool
-    @Environment(\.verticalSizeClass) private var verticalSizeClass
-
-    func body(content: Content) -> some View {
-        let edges = MobileTerminalSafeAreaExpansionPolicy.edges(
-            context: context,
-            hasCompactVerticalSize: verticalSizeClass == .compact,
-            includesBottom: includesBottom
-        )
-        if edges.hasEdges {
-            content
-                .ignoresSafeArea(.container, edges: edges.edgeSet)
-        } else {
-            content
-        }
-    }
-}
-
-extension View {
-    func mobileTerminalSafeAreaExpansion(
-        context: MobileTerminalSafeAreaContext,
-        includesBottom: Bool = true
-    ) -> some View {
-        modifier(MobileCompactLandscapeTerminalSafeAreaCompensation(
-            context: context,
-            includesBottom: includesBottom
-        ))
-    }
-}
-
-@MainActor
-private enum MobileTerminalDeviceSafeArea {
-    static var landscapeCameraEdge: MobileTerminalLandscapeCameraEdge {
-        MobileTerminalLandscapeCameraEdgeResolver.edge(for: windowOrientation)
-    }
-
-    static var bottomInset: CGFloat {
-        if let keyWindow {
-            return keyWindow.safeAreaInsets.bottom
-        }
-        return 0
-    }
-
-    static func horizontalInsets(fallback: EdgeInsets) -> EdgeInsets {
-        if let keyWindow {
-            let insets = keyWindow.safeAreaInsets
-            return EdgeInsets(
-                top: insets.top,
-                leading: insets.left,
-                bottom: insets.bottom,
-                trailing: insets.right
-            )
-        }
-        return fallback
-    }
-
-    private static var windowOrientation: MobileTerminalWindowOrientation {
-        guard let windowScene = keyWindow?.windowScene else {
-            return .unknown
-        }
-        switch windowScene.interfaceOrientation {
-        case .portrait:
-            return .portrait
-        case .portraitUpsideDown:
-            return .portraitUpsideDown
-        case .landscapeLeft:
-            return .landscapeLeft
-        case .landscapeRight:
-            return .landscapeRight
-        case .unknown:
-            return .unknown
-        @unknown default:
-            return .unknown
-        }
-    }
-
-    private static var keyWindow: UIWindow? {
-        for scene in UIApplication.shared.connectedScenes {
-            guard let windowScene = scene as? UIWindowScene else { continue }
-            if let window = windowScene.windows.first(where: \.isKeyWindow) {
-                return window
-            }
-        }
-        return nil
-    }
-}
-
-#endif
-
-extension View {
-    @ViewBuilder
-    func mobilePlainTextInput() -> some View {
-        #if os(iOS)
-        self
-            .textInputAutocapitalization(.never)
-            .autocorrectionDisabled()
-        #else
-        self
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileEmailTextInput() -> some View {
-        #if os(iOS)
-        self
-            .keyboardType(.emailAddress)
-            .textContentType(.emailAddress)
-            .textInputAutocapitalization(.never)
-            .autocorrectionDisabled()
-        #else
-        self
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileOneTimeCodeInput() -> some View {
-        #if os(iOS)
-        self
-            .keyboardType(.numberPad)
-            .textContentType(.oneTimeCode)
-        #else
-        self
-        #endif
-    }
-
-    @ViewBuilder
-    func addDeviceInputBehavior(_ kind: AddDeviceInputKind) -> some View {
-        #if os(iOS)
-        switch kind {
-        case .text:
-            self
-                .textInputAutocapitalization(.words)
-                .autocorrectionDisabled()
-        case .url:
-            self
-                .keyboardType(.URL)
-                .textContentType(.URL)
-                .textInputAutocapitalization(.never)
-                .autocorrectionDisabled()
-        case .number:
-            self
-                .keyboardType(.numberPad)
-                .textInputAutocapitalization(.never)
-                .autocorrectionDisabled()
-        }
-        #else
-        self
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileInlineNavigationTitle() -> some View {
-        #if os(iOS)
-        self.navigationBarTitleDisplayMode(.inline)
-        #else
-        self
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileTerminalNavigationChrome() -> some View {
-        #if os(iOS)
-        self
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarBackground(TerminalPalette.background, for: .navigationBar)
-            .toolbarBackground(.visible, for: .navigationBar)
-            .toolbarColorScheme(.dark, for: .navigationBar)
-        #else
-        self
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileGlassButton() -> some View {
-        #if os(iOS)
-        if #available(iOS 26.0, *) {
-            self
-                .buttonStyle(.glass)
-                .buttonBorderShape(.capsule)
-                .controlSize(.extraLarge)
-        } else {
-            self
-                .buttonStyle(.bordered)
-                .controlSize(.large)
-        }
-        #else
-        self
-            .buttonStyle(.bordered)
-            .controlSize(.large)
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileGlassProminentButton() -> some View {
-        #if os(iOS)
-        if #available(iOS 26.0, *) {
-            self
-                .buttonStyle(.glassProminent)
-                .buttonBorderShape(.capsule)
-                .controlSize(.extraLarge)
-        } else {
-            self
-                .buttonStyle(.borderedProminent)
-                .controlSize(.large)
-        }
-        #else
-        self
-            .buttonStyle(.borderedProminent)
-            .controlSize(.large)
-        #endif
-    }
-
-    @ViewBuilder
-    func mobileGlassPill() -> some View {
-        #if os(iOS)
-        if #available(iOS 26.0, *) {
-            self.glassEffect(.regular.interactive(), in: .capsule)
-        } else {
-            self
-                .background(.thinMaterial, in: Capsule())
-                .overlay(Capsule().stroke(.white.opacity(0.18), lineWidth: 1))
-        }
-        #else
-        self
-            .background(.thinMaterial, in: Capsule())
-            .overlay(Capsule().stroke(.white.opacity(0.18), lineWidth: 1))
-        #endif
-    }
-}
