+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/PairingView.swift
@@ -305,136 +305,3 @@ private enum AddDeviceField: Hashable {
     case host
     case port
 }
-
-enum AddDeviceInputKind {
-    case text
-    case url
-    case number
-}
-
-/// Whether a scanned QR payload is a cmux pairing link.
-///
-/// cmux pairing QR codes carry a `cmux-ios://` deep link; any other QR content
-/// (a website URL, a Wi-Fi join code) must be ignored so the scanner never hands
-/// the connection layer a non-pairing string.
-func mobilePairingScannerAcceptsCode(_ code: String) -> Bool {
-    code.hasPrefix("cmux-ios://")
-}
-
-#if os(iOS)
-struct MobilePairingScannerSheet: View {
-    let onCode: (String) -> Void
-    @Environment(\.dismiss) private var dismiss
-    private let authorization = CameraAuthorization()
-    @State private var authorizationStatus = CameraAuthorization().videoStatus
-
-    var body: some View {
-        NavigationStack {
-            Group {
-                switch authorizationStatus {
-                case .authorized:
-                    QRCodeScannerView { code in
-                        onCode(code)
-                    }
-                    .ignoresSafeArea(edges: .bottom)
-                case .notDetermined:
-                    ProgressView()
-                        .task {
-                            authorizationStatus = await authorization.requestVideoAccess()
-                        }
-                case .denied, .restricted:
-                    ContentUnavailableView(
-                        L10n.string("mobile.pairing.cameraDenied", defaultValue: "Camera Access Required"),
-                        systemImage: "camera.fill"
-                    )
-                @unknown default:
-                    ContentUnavailableView(
-                        L10n.string("mobile.pairing.cameraUnavailable", defaultValue: "Camera Unavailable"),
-                        systemImage: "camera.fill"
-                    )
-                }
-            }
-            .navigationTitle(L10n.string("mobile.pairing.scannerTitle", defaultValue: "Scan QR Code"))
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .topBarTrailing) {
-                    Button {
-                        dismiss()
-                    } label: {
-                        Text(L10n.string("mobile.pairing.scannerCancel", defaultValue: "Cancel"))
-                    }
-                    .accessibilityIdentifier("MobileScannerCancelButton")
-                }
-            }
-        }
-    }
-}
-
-/// SwiftUI host for the ``CmuxMobileCamera`` QR-capture controller.
-///
-/// Owns one ``QRCodeScanStream`` per presentation, mounts the package's
-/// ``QRCodeCaptureController``, and forwards accepted `cmux-ios://` codes to
-/// `onCode`. The AVCaptureSession lifecycle now lives entirely in the camera
-/// service; this wrapper only bridges the stream to a SwiftUI callback.
-struct QRCodeScannerView: UIViewControllerRepresentable {
-    let onCode: (String) -> Void
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(onCode: onCode)
-    }
-
-    func makeUIViewController(context: Context) -> QRCodeCaptureController {
-        let stream = QRCodeScanStream()
-        context.coordinator.observe(stream: stream)
-        return QRCodeCaptureController(
-            stream: stream,
-            accepts: mobilePairingScannerAcceptsCode,
-            unavailableText: L10n.string("mobile.pairing.cameraUnavailable", defaultValue: "Camera Unavailable")
-        )
-    }
-
-    func updateUIViewController(_ uiViewController: QRCodeCaptureController, context: Context) {
-        context.coordinator.onCode = onCode
-    }
-
-    static func dismantleUIViewController(_ uiViewController: QRCodeCaptureController, coordinator: Coordinator) {
-        coordinator.cancel()
-    }
-
-    @MainActor
-    final class Coordinator {
-        var onCode: (String) -> Void
-        private var task: Task<Void, Never>?
-
-        init(onCode: @escaping (String) -> Void) {
-            self.onCode = onCode
-        }
-
-        func observe(stream: QRCodeScanStream) {
-            task?.cancel()
-            task = Task { @MainActor [weak self] in
-                for await code in stream.codes {
-                    guard !Task.isCancelled else { return }
-                    self?.onCode(code)
-                }
-            }
-        }
-
-        func cancel() {
-            task?.cancel()
-            task = nil
-        }
-    }
-}
-#else
-struct MobilePairingScannerSheet: View {
-    let onCode: (String) -> Void
-
-    var body: some View {
-        ContentUnavailableView(
-            L10n.string("mobile.pairing.cameraUnavailable", defaultValue: "Camera Unavailable"),
-            systemImage: "camera.fill"
-        )
-    }
-}
-#endif
