+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/View+MobileGlass.swift
@@ -1,0 +1,63 @@
+import SwiftUI
+
+extension View {
+    /// Glass (iOS 26+) or bordered button styling for secondary sign-in actions.
+    @ViewBuilder
+    func mobileGlassButton() -> some View {
+        #if os(iOS)
+        if #available(iOS 26.0, *) {
+            self
+                .buttonStyle(.glass)
+                .buttonBorderShape(.capsule)
+                .controlSize(.extraLarge)
+        } else {
+            self
+                .buttonStyle(.bordered)
+                .controlSize(.large)
+        }
+        #else
+        self
+            .buttonStyle(.bordered)
+            .controlSize(.large)
+        #endif
+    }
+
+    /// Prominent glass (iOS 26+) or bordered-prominent primary button styling.
+    @ViewBuilder
+    func mobileGlassProminentButton() -> some View {
+        #if os(iOS)
+        if #available(iOS 26.0, *) {
+            self
+                .buttonStyle(.glassProminent)
+                .buttonBorderShape(.capsule)
+                .controlSize(.extraLarge)
+        } else {
+            self
+                .buttonStyle(.borderedProminent)
+                .controlSize(.large)
+        }
+        #else
+        self
+            .buttonStyle(.borderedProminent)
+            .controlSize(.large)
+        #endif
+    }
+
+    /// Glass (iOS 26+) or thin-material capsule pill background for input fields.
+    @ViewBuilder
+    func mobileGlassPill() -> some View {
+        #if os(iOS)
+        if #available(iOS 26.0, *) {
+            self.glassEffect(.regular.interactive(), in: .capsule)
+        } else {
+            self
+                .background(.thinMaterial, in: Capsule())
+                .overlay(Capsule().stroke(.white.opacity(0.18), lineWidth: 1))
+        }
+        #else
+        self
+            .background(.thinMaterial, in: Capsule())
+            .overlay(Capsule().stroke(.white.opacity(0.18), lineWidth: 1))
+        #endif
+    }
+}
