+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceDetailContainer.swift
@@ -1,0 +1,60 @@
+import Foundation
+import CMUXMobileCore
+import CmuxMobileDiagnostics
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct WorkspaceDetailContainer: View {
+    @Bindable var store: CMUXMobileShellStore
+    let workspaceID: MobileWorkspacePreview.ID?
+    let createWorkspace: () -> Void
+    let safeAreaContext: MobileTerminalSafeAreaContext
+
+    private var workspace: MobileWorkspacePreview? {
+        if let workspaceID {
+            return store.workspaces.first { $0.id == workspaceID } ?? store.selectedWorkspace
+        }
+        return store.selectedWorkspace
+    }
+
+    var body: some View {
+        if let workspace {
+            WorkspaceDetailView(
+                host: store.connectedHostName,
+                workspace: workspace,
+                store: store,
+                selectedTerminalID: Binding(
+                    get: { store.selectedTerminalID },
+                    set: { store.selectTerminal($0) }
+                ),
+                createWorkspace: createWorkspace,
+                createTerminal: store.createTerminal,
+                reportTerminalViewport: store.reportTerminalViewport,
+                sendTerminalInput: store.sendTerminalRawInput,
+                safeAreaContext: safeAreaContext
+            )
+            .onAppear {
+                if store.selectedWorkspaceID != workspace.id {
+                    store.selectedWorkspaceID = workspace.id
+                }
+            }
+            .task(id: workspace.id) {
+                await store.openWorkspace(workspace.id)
+            }
+        } else {
+            ContentUnavailableView(
+                L10n.string("mobile.workspace.emptyTitle", defaultValue: "No Workspace"),
+                systemImage: "rectangle.stack"
+            )
+        }
+    }
+}
