+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceDetailView.swift
@@ -1,0 +1,229 @@
+import Foundation
+import CMUXMobileCore
+import CmuxMobileDiagnostics
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct WorkspaceDetailView: View {
+    let host: String
+    let workspace: MobileWorkspacePreview
+    @Bindable var store: CMUXMobileShellStore
+    @Binding var selectedTerminalID: MobileTerminalPreview.ID?
+    let createWorkspace: () -> Void
+    let createTerminal: () -> Void
+    let reportTerminalViewport: (MobileWorkspacePreview.ID, MobileTerminalPreview.ID, MobileTerminalViewportSize) -> Void
+    let sendTerminalInput: (String) -> Void
+    let safeAreaContext: MobileTerminalSafeAreaContext
+    @State private var isTerminalPickerPresented = false
+
+    private var selectedTerminal: MobileTerminalPreview? {
+        workspace.terminals.first { $0.id == selectedTerminalID } ?? workspace.terminals.first
+    }
+
+    var body: some View {
+        detailContent()
+    }
+
+    private func detailContent() -> some View {
+        // libghostty owns the bottom input accessory bar: it ships with
+        // `GhosttySurfaceView`'s `inputAccessoryView` (see
+        // `TerminalInputAccessoryAction`). The SwiftUI bar that used to
+        // live here has been removed so the two stacked toolbars from
+        // dogfood iosfin no longer fight for the same screen edge.
+        Group {
+            #if os(iOS)
+            if let terminalID = selectedTerminal?.id.rawValue {
+                GhosttySurfaceRepresentable(
+                    surfaceID: terminalID,
+                    store: store,
+                    fontSize: MobileTerminalFontPreference.defaultSize
+                )
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
+                .background(TerminalPalette.background)
+            } else {
+                TerminalPalette.background
+                    .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
+            }
+            #else
+            TerminalPalette.background
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
+            #endif
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
+        #if os(iOS)
+        .mobileTerminalSafeAreaExpansion(
+            context: safeAreaContext,
+            includesBottom: true
+        )
+        .background {
+            TerminalPalette.background
+                .ignoresSafeArea(.container, edges: [.horizontal, .bottom])
+        }
+        #else
+        .background(TerminalPalette.background)
+        #endif
+        .navigationTitle(workspace.name)
+        .mobileTerminalNavigationChrome()
+        .toolbar {
+            #if os(iOS)
+            ToolbarItemGroup(placement: .topBarTrailing) {
+                newWorkspaceToolbarButton
+                terminalPickerToolbarButton
+            }
+            #else
+            ToolbarItem {
+                terminalToolbarButtons
+            }
+        #endif
+        }
+    }
+
+    @ViewBuilder
+    private var terminalToolbarButtons: some View {
+        newWorkspaceToolbarButton
+        terminalPickerToolbarButton
+    }
+
+    private var newWorkspaceToolbarButton: some View {
+        Button(action: createWorkspaceFromToolbar) {
+            Label(L10n.string("mobile.workspace.new", defaultValue: "New Workspace"), systemImage: "plus.square.on.square")
+                .labelStyle(.iconOnly)
+        }
+        .foregroundStyle(TerminalPalette.foreground)
+        .accessibilityIdentifier("MobileTerminalNewWorkspaceButton")
+    }
+
+    private var terminalPickerToolbarButton: some View {
+        Button {
+            dismissTerminalKeyboardForChrome()
+            isTerminalPickerPresented = true
+        } label: {
+            Label(
+                selectedTerminal?.name ?? L10n.string("mobile.terminal.select", defaultValue: "Terminal"),
+                systemImage: "terminal"
+            )
+            .labelStyle(.iconOnly)
+        }
+        .foregroundStyle(TerminalPalette.foreground)
+        .accessibilityIdentifier("MobileTerminalDropdown")
+        .accessibilityValue(host)
+        .popover(isPresented: $isTerminalPickerPresented, arrowEdge: .top) {
+            terminalPickerContent
+        }
+    }
+
+    private var terminalPickerContent: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            Text(L10n.string("mobile.terminal.picker.title", defaultValue: "Terminals"))
+                .font(.headline)
+                .padding(.horizontal, 16)
+                .padding(.top, 14)
+                .padding(.bottom, 8)
+
+            ForEach(workspace.terminals) { terminal in
+                Button {
+                    selectTerminalFromPicker(terminal.id)
+                } label: {
+                    Label(
+                        terminal.name,
+                        systemImage: terminal.id == selectedTerminal?.id ? "checkmark.circle.fill" : "terminal"
+                    )
+                    .frame(maxWidth: .infinity, alignment: .leading)
+                    .contentShape(Rectangle())
+                }
+                .buttonStyle(.plain)
+                .padding(.horizontal, 16)
+                .padding(.vertical, 10)
+                .accessibilityIdentifier("MobileTerminalMenuItem-\(terminal.id.rawValue)")
+            }
+
+            Divider()
+                .padding(.vertical, 4)
+
+            Button(action: createWorkspaceFromTerminalPicker) {
+                Label(L10n.string("mobile.workspace.new", defaultValue: "New Workspace"), systemImage: "plus.square.on.square")
+                    .frame(maxWidth: .infinity, alignment: .leading)
+                    .contentShape(Rectangle())
+            }
+            .buttonStyle(.plain)
+            .padding(.horizontal, 16)
+            .padding(.vertical, 10)
+            .accessibilityIdentifier("MobileNewWorkspaceMenuItem")
+
+            Button(action: createTerminalFromToolbar) {
+                Label(L10n.string("mobile.terminal.new", defaultValue: "New Terminal"), systemImage: "plus")
+                    .frame(maxWidth: .infinity, alignment: .leading)
+                    .contentShape(Rectangle())
+            }
+            .buttonStyle(.plain)
+            .padding(.horizontal, 16)
+            .padding(.vertical, 10)
+            .accessibilityIdentifier("MobileNewTerminalMenuItem")
+
+            #if DEBUG && canImport(UIKit)
+            Button(action: copyDebugLogsFromMenu) {
+                // DEV-only debug tooling; not shipped, so not localized.
+                Label("Copy Debug Logs", systemImage: "doc.on.clipboard")
+                    .frame(maxWidth: .infinity, alignment: .leading)
+                    .contentShape(Rectangle())
+            }
+            .buttonStyle(.plain)
+            .padding(.horizontal, 16)
+            .padding(.vertical, 10)
+            .accessibilityIdentifier("MobileCopyDebugLogsMenuItem")
+            #endif
+        }
+        .frame(minWidth: 240, maxWidth: 320, alignment: .leading)
+        .presentationCompactAdaptation(.popover)
+    }
+
+    #if DEBUG && canImport(UIKit)
+    private func copyDebugLogsFromMenu() {
+        isTerminalPickerPresented = false
+        // Include "what the user sees" (the visible terminal text) above the
+        // debug log so a pasted bug report shows the on-screen content too.
+        let terminalText = GhosttySurfaceView.visibleTerminalSnapshot()
+        Task { @MainActor in
+            let count = await MobileDebugLog.shared.copyToPasteboard(prepending: terminalText)
+            UINotificationFeedbackGenerator().notificationOccurred(.success)
+            NSLog("cmux.terminal copied %d debug log lines + visible terminal to pasteboard", count)
+        }
+    }
+    #endif
+
+    private func createWorkspaceFromToolbar() {
+        dismissTerminalKeyboardForChrome()
+        createWorkspace()
+    }
+
+    private func createWorkspaceFromTerminalPicker() {
+        dismissTerminalKeyboardForChrome()
+        isTerminalPickerPresented = false
+        createWorkspace()
+    }
+
+    private func createTerminalFromToolbar() {
+        dismissTerminalKeyboardForChrome()
+        isTerminalPickerPresented = false
+        createTerminal()
+    }
+
+    private func selectTerminalFromPicker(_ terminalID: MobileTerminalPreview.ID) {
+        dismissTerminalKeyboardForChrome()
+        isTerminalPickerPresented = false
+        selectedTerminalID = terminalID
+    }
+
+    private func dismissTerminalKeyboardForChrome() {
+        dismissMobileKeyboard()
+    }
+}
