+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceNavigationRow.swift
@@ -1,0 +1,48 @@
+import Foundation
+import CMUXMobileCore
+import CmuxMobileDiagnostics
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct WorkspaceNavigationRow: View {
+    let workspace: MobileWorkspacePreview
+    let host: String
+    let isSelected: Bool
+    let navigationStyle: WorkspaceNavigationStyle
+    let selectWorkspace: (MobileWorkspacePreview.ID) -> Void
+
+    var body: some View {
+        Group {
+            switch navigationStyle {
+            case .push:
+                NavigationLink(value: workspace.id) {
+                    WorkspaceRow(workspace: workspace, host: host, isSelected: false)
+                }
+                .simultaneousGesture(TapGesture().onEnded {
+                    selectWorkspace(workspace.id)
+                })
+            case .sidebar:
+                Button {
+                    selectWorkspace(workspace.id)
+                } label: {
+                    WorkspaceRow(workspace: workspace, host: host, isSelected: isSelected)
+                }
+                .buttonStyle(.plain)
+            }
+        }
+        .accessibilityElement(children: .combine)
+        .accessibilityAddTraits(.isButton)
+        .accessibilityIdentifier("MobileWorkspaceRow-\(workspace.id.rawValue)")
+        .accessibilityLabel(workspace.name)
+        .accessibilityValue(workspace.accessibilitySummary(host: host))
+    }
+}
