+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceNavigationStyle.swift
@@ -1,0 +1,19 @@
+import Foundation
+import CMUXMobileCore
+import CmuxMobileDiagnostics
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+enum WorkspaceNavigationStyle {
+    case push
+    case sidebar
+}
