+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceRow.swift
@@ -1,0 +1,85 @@
+import Foundation
+import CMUXMobileCore
+import CmuxMobileDiagnostics
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct WorkspaceRow: View {
+    let workspace: MobileWorkspacePreview
+    let host: String
+    let isSelected: Bool
+
+    var body: some View {
+        HStack(alignment: .top, spacing: 12) {
+            WorkspaceAvatar(workspace: workspace)
+
+            VStack(alignment: .leading, spacing: 4) {
+                HStack(alignment: .firstTextBaseline, spacing: 8) {
+                    Text(workspace.name)
+                        .font(.headline)
+                        .foregroundStyle(isSelected ? Color.accentColor : Color.primary)
+                        .lineLimit(1)
+
+                    Spacer(minLength: 8)
+
+                    Text(workspace.timestampOrStatus(host: host))
+                        .font(.subheadline)
+                        .foregroundStyle(.secondary)
+                        .lineLimit(1)
+                }
+
+                Text(workspace.previewLine)
+                    .font(.subheadline)
+                    .foregroundStyle(.secondary)
+                    .lineLimit(1)
+
+                HStack(spacing: 6) {
+                    Circle()
+                        .fill(workspace.statusColor)
+                        .frame(width: 7, height: 7)
+
+                    Text(workspace.detailLine(host: host))
+                        .font(.caption)
+                        .foregroundStyle(.secondary)
+                        .lineLimit(1)
+                }
+            }
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .padding(.vertical, 8)
+        .padding(.horizontal, isSelected ? 10 : 0)
+        .background {
+            if isSelected {
+                RoundedRectangle(cornerRadius: 14, style: .continuous)
+                    .fill(Color.accentColor.opacity(0.14))
+            }
+        }
+        .contentShape(Rectangle())
+    }
+}
+
+struct WorkspaceAvatar: View {
+    let workspace: MobileWorkspacePreview
+
+    var body: some View {
+        ZStack {
+            Circle()
+                .fill(workspace.avatarGradient)
+                .frame(width: 48, height: 48)
+
+            Image(systemName: workspace.avatarSymbolName)
+                .font(.headline)
+                .foregroundStyle(.white)
+                .accessibilityHidden(true)
+        }
+    }
+}
