+++ Packages/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceShellView.swift
@@ -1,0 +1,181 @@
+import Foundation
+import CMUXMobileCore
+import CmuxMobileDiagnostics
+import CmuxMobileShell
+import CmuxMobileShellModel
+import CmuxMobileSupport
+import CmuxMobileTerminal
+import CmuxMobileWorkspace
+import SwiftUI
+#if os(iOS)
+@preconcurrency import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+struct WorkspaceShellView: View {
+    @Bindable var store: CMUXMobileShellStore
+    let signOut: () -> Void
+    @State private var compactNavigationPath: [MobileWorkspacePreview.ID] = []
+    @State private var pendingCompactCreateNavigationWorkspaceIDs: Set<MobileWorkspacePreview.ID>?
+    @State private var hasPresentedSplitDetail = false
+    @State private var splitColumnVisibility: NavigationSplitViewVisibility = .automatic
+    #if os(iOS)
+    @Environment(\.horizontalSizeClass) private var horizontalSizeClass
+    @Environment(\.verticalSizeClass) private var verticalSizeClass
+    #endif
+
+    private var usesCompactStack: Bool {
+        #if os(iOS)
+        MobileWorkspaceShellLayoutPolicy.usesCompactStack(
+            horizontalSizeClass: horizontalSizeClass,
+            verticalSizeClass: verticalSizeClass
+        )
+        #else
+        false
+        #endif
+    }
+
+    var body: some View {
+        Group {
+            if usesCompactStack {
+                stackLayout
+            } else {
+                splitLayout
+            }
+        }
+        .onChange(of: usesCompactStack) { _, isCompact in
+            guard isCompact, hasPresentedSplitDetail, let selectedWorkspaceID = store.selectedWorkspaceID else {
+                return
+            }
+            compactNavigationPath = [selectedWorkspaceID]
+        }
+        .accessibilityIdentifier("MobileWorkspaceShell")
+        .overlay(alignment: .top) {
+            MobileConnectionRecoveryBanner(store: store)
+        }
+    }
+
+    private var stackLayout: some View {
+        NavigationStack(path: $compactNavigationPath) {
+            WorkspaceListView(
+                workspaces: store.workspaces,
+                selectedWorkspaceID: store.selectedWorkspaceID,
+                host: store.connectedHostName,
+                navigationStyle: .push,
+                selectWorkspace: selectWorkspace,
+                createWorkspace: createWorkspaceInCompactStack,
+                rescanQR: { store.disconnectAndForgetActiveMac() },
+                signOut: signOut
+            )
+            .navigationDestination(for: MobileWorkspacePreview.ID.self) { workspaceID in
+                workspaceDestination(for: workspaceID, createWorkspace: createWorkspaceInCompactStack)
+            }
+        }
+        .onChange(of: store.selectedWorkspaceID) { _, selectedWorkspaceID in
+            if let createdPath = WorkspaceShellCompactNavigationPolicy.pathForCreatedWorkspaceSelection(
+                currentPath: compactNavigationPath,
+                selectedWorkspaceID: selectedWorkspaceID,
+                existingWorkspaceIDs: pendingCompactCreateNavigationWorkspaceIDs
+            ) {
+                pendingCompactCreateNavigationWorkspaceIDs = nil
+                compactNavigationPath = createdPath
+                autoOpenSelectedWorkspaceForSoakIfNeeded()
+                return
+            }
+            compactNavigationPath = WorkspaceShellCompactNavigationPolicy.pathForSelectionChange(
+                currentPath: compactNavigationPath,
+                selectedWorkspaceID: selectedWorkspaceID
+            )
+            autoOpenSelectedWorkspaceForSoakIfNeeded()
+        }
+        .onChange(of: compactNavigationPath) { _, path in
+            guard let selectedWorkspaceID = path.last,
+                  store.selectedWorkspaceID != selectedWorkspaceID else {
+                return
+            }
+            store.selectedWorkspaceID = selectedWorkspaceID
+        }
+        .onChange(of: store.workspaces.map(\.id)) { _, workspaceIDs in
+            compactNavigationPath.removeAll { !workspaceIDs.contains($0) }
+            autoOpenSelectedWorkspaceForSoakIfNeeded()
+        }
+        .onAppear {
+            autoOpenSelectedWorkspaceForSoakIfNeeded()
+        }
+    }
+
+    private var splitLayout: some View {
+        NavigationSplitView(columnVisibility: $splitColumnVisibility) {
+            WorkspaceListView(
+                workspaces: store.workspaces,
+                selectedWorkspaceID: store.selectedWorkspaceID,
+                host: store.connectedHostName,
+                navigationStyle: .sidebar,
+                selectWorkspace: selectWorkspace,
+                createWorkspace: store.createWorkspace,
+                rescanQR: { store.disconnectAndForgetActiveMac() },
+                signOut: signOut
+            )
+            .navigationSplitViewColumnWidth(min: 320, ideal: 380, max: 440)
+        } detail: {
+            workspaceDestination(
+                for: store.selectedWorkspaceID,
+                createWorkspace: store.createWorkspace,
+                safeAreaContext: splitColumnVisibility == .detailOnly ? .fullWidth : .splitSidebarVisible
+            )
+        }
+        .navigationSplitViewStyle(.balanced)
+        .onAppear {
+            hasPresentedSplitDetail = true
+        }
+    }
+
+    private func selectWorkspace(_ id: MobileWorkspacePreview.ID) {
+        pendingCompactCreateNavigationWorkspaceIDs = nil
+        store.selectedWorkspaceID = id
+        if usesCompactStack, compactNavigationPath.last != id {
+            compactNavigationPath = [id]
+        }
+    }
+
+    private func createWorkspaceInCompactStack() {
+        let existingWorkspaceIDs = Set(store.workspaces.map(\.id))
+        pendingCompactCreateNavigationWorkspaceIDs = existingWorkspaceIDs
+        store.createWorkspace()
+        if let createdPath = WorkspaceShellCompactNavigationPolicy.pathForCreatedWorkspaceSelection(
+            currentPath: compactNavigationPath,
+            selectedWorkspaceID: store.selectedWorkspaceID,
+            existingWorkspaceIDs: existingWorkspaceIDs
+        ) {
+            pendingCompactCreateNavigationWorkspaceIDs = nil
+            compactNavigationPath = createdPath
+        }
+    }
+
+    private func autoOpenSelectedWorkspaceForSoakIfNeeded() {
+        #if DEBUG
+        guard ProcessInfo.processInfo.environment["CMUX_MOBILE_SOAK_OPEN_SELECTED_WORKSPACE"] == "1",
+              compactNavigationPath.isEmpty,
+              let selectedWorkspaceID = store.selectedWorkspaceID,
+              store.workspaces.contains(where: { $0.id == selectedWorkspaceID }) else {
+            return
+        }
+        compactNavigationPath = [selectedWorkspaceID]
+        #endif
+    }
+
+    @ViewBuilder
+    private func workspaceDestination(
+        for workspaceID: MobileWorkspacePreview.ID?,
+        createWorkspace: @escaping () -> Void,
+        safeAreaContext: MobileTerminalSafeAreaContext = .fullWidth
+    ) -> some View {
+        WorkspaceDetailContainer(
+            store: store,
+            workspaceID: workspaceID,
+            createWorkspace: createWorkspace,
+            safeAreaContext: safeAreaContext
+        )
+    }
+}
